Added Basic SwitchTube Support in Video Session
New SwitchTube WaveSurfer-like handling of interaction with iFrame embedding of video. External embedding works via URL only (video tag is created the same way the internal video upload works)

+++ apps/api/src/app/media/media.service.ts
@@ -68,6 +68,11 @@ export class MediaService {
 
   async remove(id: number) {
     const media = await this.findOne(id);
+
+    if (!media) {
+      throw new BadRequestException('Media not found');
+    }
+
     await this.deleteMediaFile(id, media.name);
 
     await this.mediaRepository.delete(id);
